requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright>=1.42.0
pydantic==2.4.2
pydantic-settings==2.1.0
emoji>=2.2.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1